# spec/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AgentAPI Proxy",
    "description": "API for managing agent sessions, schedules, and settings. AgentAPI Proxy provides a unified interface for running AI agents with Kubernetes-based session management.",
    "version": "1.0.0",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "paths": {
...
    "/schedules": {
      "post": {
        "summary": "Create a new schedule",
        "description": "Creates a new schedule for delayed or recurring session execution.",
        "operationId": "createSchedule",
        "tags": [
          "Schedules"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateScheduleRequest"
              },
              "examples": {
                "one-time": {
                  "summary": "One-time delayed execution",
                  "value": {
                    "name": "Code Review",
                    "scheduled_at": "2025-01-15T14:00:00Z",
                    "session_config": {
                      "tags": {
                        "repository": "org/repo"
                      },
                      "params": {
                        "message": "Review PRs"
                      }
                    }
                  }
                },
                "recurring": {
                  "summary": "Recurring execution (cron)",
                  "value": {
                    "name": "Daily Standup",
                    "cron_expr": "0 9 * * 1-5",
                    "session_config": {
                      "tags": {
                        "repository": "org/standup-bot"
                      },
                      "params": {
                        "message": "Generate daily standup"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Schedule created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "get": {
        "summary": "List schedules",
        "description": "Returns a list of schedules. Non-admin users can see their own schedules and team-scoped schedules they have access to.",
        "operationId": "listSchedules",
        "tags": [
          "Schedules"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by schedule status",
            "schema": {
              "$ref": "#/components/schemas/ScheduleStatus"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "description": "Filter by resource scope",
            "schema": {
              "$ref": "#/components/schemas/ResourceScope"
            }
          },
          {
            "name": "team_id",
            "in": "query",
            "description": "Filter by team ID (e.g., 'org/team-slug')",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of schedules",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "schedules": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ScheduleResponse"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/schedules/{id}": {
      "get": {
        "summary": "Get a schedule",
        "description": "Returns a specific schedule by ID.",
        "operationId": "getSchedule",
        "tags": [
          "Schedules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Schedule ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Schedule details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - can only access own schedules"
          },
          "404": {
            "description": "Schedule not found"
          }
        }
      },
      "put": {
        "summary": "Update a schedule",
        "description": "Updates an existing schedule.",
        "operationId": "updateSchedule",
        "tags": [
          "Schedules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Schedule ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateScheduleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Schedule updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "403": {
            "description": "Forbidden - can only update own schedules"
          },
          "404": {
            "description": "Schedule not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a schedule",
        "description": "Deletes a schedule by ID.",
        "operationId": "deleteSchedule",
        "tags": [
          "Schedules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Schedule ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Schedule deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Schedule deleted successfully"
                    },
                    "id": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - can only delete own schedules"
          },
          "404": {
            "description": "Schedule not found"
          }
        }
      }
    },
    "/schedules/{id}/trigger": {
      "post": {
        "summary": "Manually trigger a schedule",
        "description": "Immediately executes a schedule, creating a new session.",
        "operationId": "triggerSchedule",
        "tags": [
          "Schedules"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Schedule ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Schedule triggered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "session_id": {
                      "type": "string",
                      "description": "ID of the created or reused session"
                    },
                    "triggered_at": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "session_reused": {
                      "type": "boolean",
                      "description": "Indicates whether an existing session was reused instead of creating a new one"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - can only trigger own schedules"
          },
          "404": {
            "description": "Schedule not found"
          },
          "500": {
            "description": "Failed to create session"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateScheduleRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the schedule"
          },
          "scope": {
            "$ref": "#/components/schemas/ResourceScope",
            "description": "Ownership scope of the schedule (default: user)"
          },
          "team_id": {
            "type": "string",
            "description": "Team identifier (e.g., 'org/team-slug') when scope is 'team'"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "paused"
            ],
            "description": "Initial status of the schedule (default: active)"
          },
          "scheduled_at": {
            "type": "string",
            "format": "date-time",
            "description": "When to execute (ISO8601). Either scheduled_at or cron_expr must be set."
          },
          "cron_expr": {
            "type": "string",
            "description": "Cron expression for recurring execution (e.g., '0 9 * * 1-5')",
            "example": "0 9 * * 1-5"
          },
          "timezone": {
            "type": "string",
            "description": "Timezone for the schedule (default: Asia/Tokyo)",
            "example": "Asia/Tokyo"
          },
          "session_config": {
            "$ref": "#/components/schemas/SessionConfig"
          }
        }
      },
...
      "ScheduleStatus": {
        "type": "string",
        "enum": [
          "active",
          "paused",
          "completed"
        ],
        "description": "Schedule status"
      },
...
      "ResourceScope": {
        "type": "string",
        "enum": [
          "user",
          "team"
        ],
        "default": "user",
        "description": "Ownership scope. 'user' for personal resources, 'team' for team-shared resources"
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "GitHub OAuth token"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Static API key"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    },
    {
      "apiKey": []
    }
  ],
...
}